Moved function in an own file
